+++ code/modules/mob/living/carbon/human/human_defines.dm
@@ -139,4 +139,3 @@
 
 	///The Examine Panel TGUI.
 	var/datum/examine_panel/tgui = new() //create the datum
-
