+++ modular_tfn/code/modules/mob/living/carbon/human/human_defines.dm
@@ -1,3 +1,5 @@
 /mob/living/carbon/human
 	var/headshot_link = null
 	var/ooc_notes
+	var/flavor_text_nsfw
+	var/character_notes
